# metal : optimize multi-sequence FA vec kernel