@ggerganov 2025-05-13
@JohannesGaessler I am wondering if we can apply the same optimization to the CUDA FA kernels. I don't have currently a suitable CUDA machine that I can use for development, so just sharing some thoughts if you have some time to try out.

The idea is that the FA-vec kernel can directly skip blocks for which the KQ mask is full of -INF. For single-sequence generations this does not happen so we won't see any improvement, but it should improve a lot multi-sequence generations and the [upcoming SWA models](#13194).

I took a quick look at the code and have the following suggestion:

```diff
diff --git a/ggml/src/ggml-cuda/fattn-vec-f16.cuh b/ggml/src/ggml-cuda/fattn-vec-f16.cuh
index d96e39212..171652dc8 100644
--- a/ggml/src/ggml-cuda/fattn-vec-f16.cuh
+++ b/ggml/src/ggml-cuda/fattn-vec-f16.cuh
@@ -175,6 +175,22 @@ static __global__ void flash_attn_vec_ext_f16(
     for (int k_VKQ_0 = blockIdx.y*D; k_VKQ_0 < ne11; k_VKQ_0 += gridDim.y*D) {
         // Calculate KQ tile and keep track of new maximum KQ values:
 
+        // mask -INF blocks
+        half mask_max = 0.0h;
+#pragma unroll
+        for (int j = 0; j < ncols; ++j) {
+            for (int i_KQ_0 = 0; i_KQ_0 < D; i_KQ_0 += nwarps) {
+                const int i_KQ = i_KQ_0 + threadIdx.y;
+                // reuse the array for masking -INF blocks
+                mask_max = max(mask_max, maskh[j*ne11 + k_VKQ_0 + i_KQ]);
+            }
+        }
+
+        mask_max = warp_reduce_max(mask_max); // TODO: not sure what it the function, but simply warp-reduce the var
+        if (mask_max == -INF_HALF) {
+            continue;
+        }
+
         // For unknown reasons using a half array of size 1 for kqmax_new causes a performance regression,
         // see https://github.com/ggerganov/llama.cpp/pull/7061 .
         // Therefore this variable is defined twice but only used once (so that the compiler can optimize out the unused variable).
```

It's not tested at all, but I hope mainly to illustrate the idea. There might be a better way to do it.

The same logic should be possible to apply to the non-vec FA kernel, but this is less important for now. If you could give this a try and run the `llama-batched-bench` command from this PR (make sure to have the changes from this branch), would appreciate. I am interested in the `S_TG` numbers before and after this change. Thanks.