@JohannesGaessler 2025-05-13
For token generation the bottleneck is I/O. Tensor cores work on matrix fragments of at least 8 rows/columns. Without GQA for <= 8 tokens each value from the KV cache is being loaded only once. So for <= 8 parallel sequences you can basically never skip any KV slices unless you use an implementation without tensor cores (the vec kernels do not). The vec kernels only work for a single token, for >1 tokens they would need to use KV values for >1 Q values and you would quickly run into the same problem where each KV value is basically always used for at least one sequence and you can never skip it completely. At most you can skip some of the compute but as long as you're I/O bound all of the compute pipelines are going to be underutilized anyways.

However, if a model is using GQA then you can get a much higher arithmetic intensity for each KV value since you can apply it to multiple Q values for each sequence. In such a scenario an optimization for potentially skipping a KV slice could work because you're much less I/O bound. As it is the CUDA kernels entangle up to 64 Q columns though so for this optimization to work correctly the selection logic for the number of parallel columns needs to know that the tokens belong to different sequences.

So the bottom line is that for CUDA this optimization will only make sense for at least `8 / (Q->ne[2] / K->ne[2])` parallel sequences and only if an op param is set to indicate that there will be little to no overlap between sequences.

>I don't have currently a suitable CUDA machine that I can use for development, so just sharing some thoughts if you have some time to try out.

I have a machine with 6 RTX 4090s in my basement, I can give you access if it would be useful for your work.