@JohannesGaessler 2025-05-13
Just so I don't misunderstanding anything: multiple parallel sequences result in an increase in `Q->ne[1]`, correct?

Also I forgot: for SWA specifically modifying the kernels to skip fully masked-out KV slices would work without issue.