@JohannesGaessler 2025-05-13
The CUDA kernel in `fattn-mma-f16.cuh` is using a stream-k decomposition where it's already working on fractional tiles. If the sequences had contiguous KV cache slices you could in principle also just adjust the lower and upper bounds of those fractional tiles.