@JohannesGaessler 2025-05-13
>Edit: Btw, it might be worth trying to force using the ncols == 1 kernel because otherwise the check will never trigger for this particular test. Probably like this if I am following the logic correctly:

I should have probably been more explicit about this: you do not gain any advantage in terms of I/O if you make this change. At most you will gain an advantage in terms of compute. The vec kernels would first need to receive the same GQA optimizations that the kernel in `fattn-mma-f16.cuh` has.