@ggerganov 2025-05-13
> Just so I don't misunderstanding anything: multiple parallel sequences result in an increase in Q->ne[1], correct?

Generally yes, but there is also another use case. For example in https://github.com/ggml-org/llama.cpp/issues/10860, using the provided Python runner script in sequential mode, would have `Q->ne[1] == 1` during the text generation phase, but the different requests will be processed as different sequences (because the server has `-np 15`). So the KV cache will be very large after processing a few sequences, but the batch size will remain 1. The optimization in this PR skips a large amount of the computation in the FA vec kernel and improves the total runtime by ~20% for the Metal backend.

The script there also allows to run in parallel mode - in that case yes, `Q->ne[1] = n_parallel` during text generation.

> I have a machine with 6 RTX 4090s in my basement, I can give you access if it would be useful for your work.

No worries, I was looking for a quick test.

Just to understand, is the `flash_attn_vec_ext_f16` kernel ever executed for Ampere cards? And if yes, for which batch sizes (i.e. values of `Q->ne[1]`)?