@JohannesGaessler 2025-05-13
For Ampere that kernel should only be used for batch size 1 and if not using GQA, but only because the mma implementation turned out to be faster due to the GQA optimization.